Validate + Verify + Blacklist on refresh
When refreshing a token it must first validate and verify the old one.
If all goes ok, it should blacklist the old token before returning the
new one.

Validation excludes the exp validator, since we can refresh an expired
token (that's the whole point), but only withing a certain time lapse
(the TTR).

Validating and verifying a token also blacklists it if something fails
and then throws whatever exception was caught again.

Validators (custom or built-in) can be excluded when refreshing, but
probably more useful for custom ones if the user needs it.